Added indexing for characteristic properties
In BGS Beta2-r1531 the characeristic properties are in an array/list.  To
reference a property, you have to use the index, like this:

characteristic.Properties[0].Read

Also, changed inner double quotes to single quotes to help with clarity.

Lastly, BGS Beta2-r1531 has a bug which prevents you from adding a value,
so that known issue (with the template file) will have to wait to be
worked on until the SIG releases an update.